graphics: rename objects.h -> graphics_types.h; split vulkan_types.h
- Public POD descriptors + GPU-object handles now live in
  include/gecko/graphics/graphics_types.h (was objects.h). Name is more
  descriptive; sits alongside gecko/core/types.h and matches the rest
  of the 'types' headers in the project.
- Vulkan-internal GPU-object payloads (VulkanSwapchainData /
  VulkanTextureData / VulkanRTData / VulkanBufferData /
  VulkanPipelineData) move out of src/graphics/vulkan/vulkan_device.h
  into a new src/graphics/vulkan/vulkan_types.h alongside the
  vk_mem_alloc.h include. vulkan_device.h now only declares the
  VulkanDevice class.
- Rename test/graphics/test_objects.cpp -> test_graphics_types.cpp.
- Update docs/graphics.md + every internal include site.
