+++ test/graphics/test_graphics_types.cpp
@@ -1,4 +1,4 @@
-#include "gecko/graphics/objects.h"
+#include "gecko/graphics/graphics_types.h"
 
 #include <catch2/catch_test_macros.hpp>
